Implement the CoreSpotlightSearch implementation
On load of the region list we index the region in CoreSpotlight.
When the user searches they see a nice little UI with a
very basic description of the wine region and the title.
Tapping the UI will set the region matching the UUID.
Still need to do some combine magic to essentially queue
up the UUID when the region tree has not yet been loaded

re #32